Add Post for syslog-ng
Add a tutorial related to the installing, configuring, and securing the
syslog-ng product, including incorporating a certificate authority and
management of a central syslog-ng instance. This post will likely
contribute to a larger initiative for a data streaming pipeline that
incorporates the ability to send logs via the native and/or syslog-ng
logging solution as an alternative to agent-based log forwarding
requiring larger footprint Java agents and the like.