Add Post on Remote Busy/Free Controller
Add a post detailing how to construct a remote busy/free circuit using
ESP8266-01, RFID reader, LEDs and a bunch of supplemental components,
all in a wooden box. This device can be used to remotely control
busy/free status so someone can be informed of whether you are free or
involved in something that requires your attention.